+++ tests/testthat/test-pip.R
@@ -111,18 +111,24 @@ test_that("year selection is working", {
     fill_gaps = TRUE,
     lkup = lkup
   )
-  check <- max(lkup$ref_lkup$reporting_year)
+  check <- max(lkup$ref_lkup[country_code == "AGO"]$reporting_year)
   expect_equal(tmp$reporting_year, check)
 
   # Most recent year for all countries
+  # Should return the most recent for each country
+  # Therefore we expect having more than one year in the response
+  # Not a great unit test... be cause it will not be always true.
+  # The possibility exists that all countries will have the same maximum
+  # reporting year?
+  # To be improved
   tmp <- pip(
     country = "all",
     year = "mrv",
     povline = 1.9,
     lkup = lkup
   )
-  check <- max(lkup$svy_lkup$reporting_year)
-  expect_equal(unique(tmp$reporting_year), check)
+
+  expect_true(length(unique(tmp$reporting_year)) > 1)
 
 })
 
@@ -337,7 +343,7 @@ test_that("pop_share option is working", {
   expect_equal(nrow(tmp), 1)
 })
 
-test_that("pop_share option is returning consisten results for single microdata distributions", {
+test_that("pop_share option is returning consistent results for single microdata distributions", {
   # Average poverty line
   povline <- 2.0
 
@@ -400,7 +406,7 @@ test_that("pop_share option is returning consisten results for single microdata 
   expect_equal(povline, round(ps$poverty_line, 0))
 })
 
-test_that("pop_share option is returning consisten results for single grouped distributions", {
+test_that("pop_share option is returning consistent results for single grouped distributions", {
   # Average poverty line
   povline <- 2.0
   country <- "MNG"
@@ -465,7 +471,7 @@ test_that("pop_share option is returning consisten results for single grouped di
   expect_equal(povline, round(ps$poverty_line, 0))
 })
 
-test_that("pop_share option is returning consisten results for single aggregate distributions", {
+test_that("pop_share option is returning consistent results for single aggregate distributions", {
   skip("popshare not working for aggregate distributions")
   # Average poverty line
   povline <- 2.0
@@ -535,6 +541,35 @@ test_that("pop_share option is returning consisten results for single aggregate 
 
   expect_equal(round(pl$headcount, 2), round(ps$headcount, 2))
   expect_equal(povline, round(ps$poverty_line, 0))
+})
+
+test_that("pop_share option is disabled for aggregate distributions", {
+  # popshare is currently not working with aggregate distribution and has been
+  # disabled
+
+  povline <- 2.0
+  country <- "CHN"
+  year <- 2018
+
+  pl <- pip(
+    country = country,
+    year = year,
+    povline = povline,
+    reporting_level = "national",
+    lkup = lkup
+  )
+
+  ps <- pip(
+    country = "CHN",
+    year = 2018,
+    popshare = .5,
+    reporting_level = "national",
+    lkup = lkup
+  )
+
+  expect_equal(nrow(pl), 1)
+  expect_equal(nrow(ps), 0)
+  expect_equal(pl$distribution_type, "aggregate")
 })
 
 #Check pip country name case insensitive