+++ tests/testthat/test-utils.R
@@ -114,3 +114,218 @@ test_that("subset_lkup correctly selects countries and regions", {
   expect_true(all(country_selection %in% (unique(tmp$country_code))))
 })
 
+# select_country() test suite
+test_that("select_country works for complete country selection", {
+
+  expected_countries <- nrow(ref_lkup)
+  keep <- rep(TRUE, expected_countries)
+
+  keep <- select_country(ref_lkup, keep, "all")
+  expect_equal(length(keep), expected_countries)
+  expect_equal(all(keep), TRUE)
+
+  keep <- select_country(ref_lkup, keep, "WLD")
+  expect_equal(length(keep), expected_countries)
+  expect_equal(all(keep), TRUE)
+
+  keep <- select_country(ref_lkup, keep, c("WLD", "COL"))
+  expect_equal(length(keep), expected_countries)
+  expect_equal(all(keep), TRUE)
+})
+
+test_that("select_country works for region selection", {
+
+  region <- "SSA"
+  expected_countries <- nrow(ref_lkup[ref_lkup$region_code == region, ])
+  keep <- rep(TRUE, nrow(ref_lkup))
+
+  keep <- select_country(ref_lkup, keep, region, valid_regions = valid_regions)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_countries)
+  expect_equal(unique(ref_lkup$pcn_region_code[keep]), region)
+})
+
+test_that("select_country works for country selection", {
+
+  country <- "COL"
+  expected_countries <- nrow(ref_lkup[ref_lkup$country_code == country, ])
+  keep <- rep(TRUE, nrow(ref_lkup))
+
+  keep <- select_country(ref_lkup, keep, country, valid_regions = valid_regions)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_countries)
+  expect_equal(unique(ref_lkup$country_code[keep]), country)
+
+  country <- c("COL", "YEM", "ZMB")
+  expected_countries <- nrow(ref_lkup[ref_lkup$country_code %in% country, ])
+  keep <- rep(TRUE, nrow(ref_lkup))
+
+  keep <- select_country(ref_lkup, keep, country, valid_regions = valid_regions)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_countries)
+  expect_equal(sort(unique(ref_lkup$country_code[keep])), country)
+})
+
+test_that("select_country works for country & region selection", {
+
+  country <- "COL"
+  region <- "SSA"
+  row_keep <- ref_lkup$country_code %in% country | ref_lkup$region_code %in% region
+  expected_countries <- nrow(ref_lkup[row_keep, ])
+  keep <- rep(TRUE, nrow(ref_lkup))
+
+  keep <- select_country(ref_lkup, keep, c(country, region), valid_regions = valid_regions)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_countries)
+  expect_true(all(country %in% unique(ref_lkup$country_code[keep])))
+  expect_true(all(region %in% unique(ref_lkup$region_code[keep])))
+})
+
+# select_years() test suite
+test_that("select_years works for most recent value", {
+  # Single country
+  country <- "BFA"
+  year <- "mrv"
+  keep <- rep(TRUE, nrow(ref_lkup))
+  keep_country <- select_country(ref_lkup, keep, country, valid_regions = valid_regions)
+  tmp <- ref_lkup[keep_country, ]
+  mrv_year <- max(tmp$reporting_year)
+  expected_countries <- nrow(tmp[tmp$reporting_year == mrv_year, ])
+
+  keep <- select_years(ref_lkup, keep_country, year, country = country)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_countries)
+  expect_equal(unique(ref_lkup$reporting_year[keep]), mrv_year)
+
+  # Multiple countries
+  country <- c("BFA", "CAN")
+  year <- "mrv"
+  keep <- rep(TRUE, nrow(ref_lkup))
+  keep_country <- select_country(ref_lkup, keep, country, valid_regions = valid_regions)
+  tmp <- ref_lkup[keep_country, ]
+  mrv_year <- tmp[tmp$country_code %in% country,
+                  .SD[which.max(reporting_year)],
+                  by = country_code]$reporting_year
+  expected_countries <- nrow(tmp[tmp$reporting_year %in% mrv_year, ])
+
+  keep <- select_years(ref_lkup, keep_country, year, country = country)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_countries)
+  expect_equal(sort(unique(ref_lkup$reporting_year[keep])), sort(mrv_year))
+
+  # All countries
+  country <- "all"
+  year <- "mrv"
+  keep <- rep(TRUE, nrow(ref_lkup))
+  keep_country <- select_country(ref_lkup, keep, country, valid_regions = valid_regions)
+  tmp <- ref_lkup[keep_country, ]
+  mrv_year <- tmp[, .SD[which.max(reporting_year)],
+                  by = country_code]$reporting_year
+  expected_countries <- length(unique(ref_lkup$country_code)) # Here we expect a single year to be returned for each single country
+
+  keep <- select_years(ref_lkup, keep_country, year, country = country)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_countries)
+  expect_equal(sort(ref_lkup$reporting_year[keep]), sort(mrv_year))
+})
+
+test_that("select_years works for all year", {
+  # Single country
+  country <- "BFA"
+  year <- "all"
+  keep <- rep(TRUE, nrow(ref_lkup))
+  keep_country <- select_country(ref_lkup, keep, country, valid_regions = valid_regions)
+
+  keep <- select_years(ref_lkup, keep_country, year, country = country)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), sum(keep_country))
+
+  # Multiple countries
+  country <- c("BFA", "CAN")
+  year <- "all"
+  keep <- rep(TRUE, nrow(ref_lkup))
+  keep_country <- select_country(ref_lkup, keep, country, valid_regions = valid_regions)
+
+  keep <- select_years(ref_lkup, keep_country, year, country = country)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), sum(keep_country))
+
+  # All countries
+  country <- "all"
+  year <- "all"
+  keep <- rep(TRUE, nrow(ref_lkup))
+  keep_country <- select_country(ref_lkup, keep, country, valid_regions = valid_regions)
+
+  keep <- select_years(ref_lkup, keep_country, year, country = country)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), sum(keep_country))
+})
+
+test_that("select_years works for specific year selections", {
+  # Single year
+  country <- "BFA"
+  year <- 2008
+  keep <- rep(TRUE, nrow(ref_lkup))
+  keep_country <- select_country(ref_lkup, keep, country, valid_regions = valid_regions)
+  tmp <- ref_lkup[keep_country, ]
+  expected_countries <- nrow(tmp[tmp$reporting_year %in% year, ])
+
+  keep <- select_years(ref_lkup, keep_country, year, country = country)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_countries)
+  expect_equal(unique(ref_lkup$reporting_year[keep]), year)
+
+  # Multiple years
+  country <- c("BFA", "CAN")
+  year <- c(2008, 2010, 2018)
+  keep <- rep(TRUE, nrow(ref_lkup))
+  keep_country <- select_country(ref_lkup, keep, country, valid_regions = valid_regions)
+  tmp <- ref_lkup[keep_country, ]
+  expected_countries <- nrow(tmp[tmp$reporting_year %in% year, ])
+
+  keep <- select_years(ref_lkup, keep_country, year, country = country)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_countries)
+  expect_equal(sort(unique(ref_lkup$reporting_year[keep])), sort(year))
+
+  # Multiple years + MRV
+  # We expect all countries with their MRV year + all country/years not part of MRV query
+  # This not currently working and specific are being dropped
+  # This test is expected to fail once the correct behavior is implemented
+  country <- "all"
+  not_mrv_year <- "1991"
+  year <- c("mrv", not_mrv_year)
+  keep <- rep(TRUE, nrow(ref_lkup))
+  mrv_year <- ref_lkup[, .SD[which.max(reporting_year)],
+                       by = country_code]$reporting_year
+  expected_row_mrv <- length(unique(ref_lkup$country_code)) # Here we expect a single year to be returned for each single country
+  expected_row_not_mrv <- nrow(ref_lkup[ref_lkup$reporting_year == not_mrv_year, ])
+
+  keep <- select_years(ref_lkup, keep, year, country = country)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_row_mrv)
+  expect_equal(sort(unique(ref_lkup$reporting_year[keep])),
+               sort(unique(mrv_year)))
+})
+
+skip("Specific year selections are dropped when MRV is selected")
+test_that("select_years works for MRV + specific year selections", {
+
+  # Multiple years + MRV
+  # We expect all countries with their MRV year + all country/years not part of MRV query
+  country <- "all"
+  not_mrv_year <- "1991"
+  year <- c("mrv", not_mrv_year)
+  keep <- rep(TRUE, nrow(ref_lkup))
+  mrv_year <- ref_lkup[, .SD[which.max(reporting_year)],
+                       by = country_code]$reporting_year
+  expected_row_mrv <- length(unique(ref_lkup$country_code)) # Here we expect a single year to be returned for each single country
+  expected_row_not_mrv <- nrow(ref_lkup[ref_lkup$reporting_year == not_mrv_year, ])
+
+  keep <- select_years(ref_lkup, keep, year, country = country)
+  expect_equal(length(keep), nrow(ref_lkup))
+  expect_equal(sum(keep), expected_row_mrv + expected_row_not_mrv)
+  expect_equal(sort(unique(ref_lkup$reporting_year[keep])),
+               sort(unique(c(mrv_year, as.numeric(not_mrv_year)))))
+})
+